% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthropogenic_deposition.R
\name{A_rates}
\alias{A_rates}
\title{Simulate anthropogenic deposition}
\usage{
A_rates(
  x,
  kcalpers = 2,
  kcalmeat_eat = 0.45,
  kcalmeat_prod = 1.5,
  in_camp_eat = 0.55,
  in_camp_stay = 13,
  kg = 0.07
)
}
\arguments{
\item{x}{Integer (user provided), vector or data.frame. If integer, it is the number of
people inhabiting the site. If data.frame, the number of people is the number
of rows. If vector, it is the length of the vector. It is \eqn{Pop(t)} in equation 3}

\item{kcalpers}{Quantity of kilocalories consumed per day per adult person.
It corresponds to \eqn{B(t)} in the equation 3. But in the formula it is per year.
Tt has a range of \verb{[1.5,2.5]}. Default is 2.}

\item{kcalmeat_eat}{Proportion of kilocalories extracted from meat. Range \link{0,1}.
Default is 0.45, based on Cordain et al (2000). It is \eqn{M(t)} in equation 3.}

\item{kcalmeat_prod}{Quantity of kiocalories per meat kilogram. Range \link{1,2.5}
Default is 1.5, considering goat meat. It is \eqn{R(t)} in euqation 3.}

\item{in_camp_eat}{Proportion of food consumed within the camp. Range \link{0,1}.
Default is 0.55 based on Collette-Barbesque et al. (2016). \eqn{S(t)} in equation 3.}

\item{in_camp_stay}{Proportion of time spent in a specific camp. Valid for
groups with high mobility. The proportion is computed within the function, but
the user introduces the weeks of occupation of the camp, where the maximum is
52 (full year). Default is 13 (weeks, or 0.25 of the year, or three months a year).
It corresponds to \eqn{O(t)} in equation 3.}

\item{kg}{Bone proportion for each animal consumed. Default is 0.07
based on Johnston et al. (2021). It corresponds to \eqn{F(t)} in equation 3.}
}
\value{
The number of samples
}
\description{
Simulation of samples generated per year (anthropogenic deposition rates)
Returns the Kilograms of bone produced per year in a site.
}
