% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{changePopSize}
\alias{changePopSize}
\title{Change population sizes}
\usage{
changePopSize(
  loosingPop,
  size,
  winingPop = NULL,
  new = F,
  method = "random",
  probs = dnorm,
  prob.option = list(sd = 10, mean = 22)
)
}
\arguments{
\item{loosingPop}{data.frame of the population that's decreasing}

\item{size}{Number of deaths or size of the population swap}

\item{winingPop}{data.frame of the population that's increasing}

\item{new}{Deprecated}

\item{method}{Currently only supports "random"}

\item{probs}{Density distribution function to calculate probabilities of individuals
to be sampled based on age}

\item{prob.option}{Options for probs functions (e.g. \code{mean} and \code{sd} for \code{dnorm})}
}
\value{
Either the updated loosing population data.frame or a list with both
the winning and loosing populations
}
\description{
It changes population sizes due to killing or swap between
a loosing and winning population
}
