% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_deposition.R
\name{D_along}
\alias{D_along}
\title{Depth protocol deposition}
\usage{
D_along(
  W_t,
  r,
  max_bone_thickness = c("m", "s", "l", "vl"),
  prop_buried = 0.9999
)
}
\arguments{
\item{W_t}{Integer (user provided), vector or data.frame. It is the number samples
produced at a specific 't'.}

\item{r}{Is the natural deposition rates. At this moment. Values greater than 0.5 are not accepted.
If values with two or more decimals are provided, the function will automatically round
the value to one decimal.}

\item{max_bone_thickness}{Maximum thickness of bones within the assemblage. Four
values are possible: \code{'s'} (small) = 2.5 cm; \code{'m'} (medium) = 5 cm; \code{'l'} (large) = 10 cm
and \code{'vl'} (very large) = 20 cm. Default is \code{'m'}.}

\item{prop_buried}{Proportion of samples buried sample at tmax, considering error. Pb
needs to be smaller than 1. Default is 0.9999, which stands for 99.99\%.
This is \eqn{\theta_{\epsilon}} in equation 9}
}
\value{
A vector of samples buried at times from tl to tm
}
\description{
It distributes the samples produced in one specific year along the depth of the
site, without any kind of post-depositional alteration, and according to
pre-established post-deposition rates. Returns a vector with the samples exponentially
distributed. The vector is as long as L/r and the error (prop_buried, \eqn{\theta_{\epsilon}}) is considered.
}
