% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{death}
\alias{death}
\title{Simulation of death process}
\usage{
death(
  x,
  pd = data.frame(Age = c(0:99), P_d = c(rep(0.14, 1), rep(0.16, 4), rep(0.05, 5),
    rep(0.01, 24), rep(0.03, 14), rep(0.1, 10), rep(0.3, 42)))
)
}
\arguments{
\item{x}{A vector or data.frame with a single row from the population matrix.
It must contain two values or columns, Age and Sex}

\item{pd}{The probability matrix for mortality by age
The age-structured data frame is based on Gurven, Kaplan and Supa, 2007.
It is extracted adapted after computation from their text (not graphs or tables)}
}
\value{
A value (0,1) where 0 = person lives and 1 = person dies, based
on pd (the probability matrix)
}
\description{
For a single individual, returns whether it lives (0) or dies (1)
Meant to be used with apply on the population data.frame x
}
