% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{environment}
\alias{environment}
\title{Environment generator}
\usage{
environment(N, omega, delta, vt = NULL)
}
\arguments{
\item{N}{the length of the timeserie to generate}

\item{omega}{the slope of the power distribution (called omega in whitehead)}

\item{delta}{the standard deviation of the global environmental fluctuation}

\item{vt}{if not NULL, the  mean of the optimum increase at a rate vt}
}
\value{
a list of N optima theta
}
\description{
Return a list of optimum as a \eqn{1/f_{noise}} of \eqn{N} steps with \eqn{sd = \delta} and
the slope of its spectrum decompoistion = -\eqn{\omega} + possibility to increase
the mean of the environment at a rate vt (in wich case the slop may not be -omega)
}
