% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extractDates}
\alias{extractDates}
\title{Extract dates}
\usage{
extractDates(record, years = NULL, n)
}
\arguments{
\item{record}{a full archaeological record (each line some depth, column represent year)}

\item{years}{better not to use, but if used should  be equal to \code{ncol(record)}}

\item{n}{final amount of sample needed}
}
\description{
Extract the dates from an archaeological record
}
