% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{fightbetterloss}
\alias{fightbetterloss}
\title{Model fight with better probabilities}
\usage{
fightbetterloss(Ne, a, b)
}
\arguments{
\item{Ne}{list of population sizes for the fighting settlement}

\item{a}{indice of the first settlement}

\item{b}{indice of the second settlement}
}
\value{
Returns the updated population size of both settlements engaged in the fight
}
\description{
A function to compute lost during a fighting.
The winner and looser are decided probabilistically from their relative size.
It then uses a binomial to model the loss of population sizes,
with probability based on their relative sizes.
}
