% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{Gpd}
\alias{Gpd}
\title{Generation of population dynamics}
\usage{
Gpd(
  x,
  K,
  W_fert_age = c(10, 45),
  M_fert_age = c(15, 55),
  p_offspring = 0.3,
  prob = 0.8,
  ...
)
}
\arguments{
\item{x}{Input data with initial population matrix. A data frame or matrix
with two columns and nrow = Initial population. One row per individual.
The first column is the age of the individual.
The second column is the sex of the individual, and must be c("F","M").
The columns must be named Age and Sex respectively.}

\item{K}{Carrying capacity.}

\item{W_fert_age}{Vector with two values. The first value is the youngest age
at which is considered that women can have children for prehistoric societies.
The second value is the oldest age at which is considered that women can
have children. Default is c(10,30).}

\item{M_fert_age}{Vector with two values. The first value is the youngest age
at which is considered that men can have children for prehistoric societies.
The second value is the oldest age at which is considered that men can have
children. Default is c(15,40)}

\item{prob}{Probability that an individual will die per year if total population
exceeds K. Default is 0.8}

\item{...}{Arguments passed to \code{\link[=death]{death()}}. The mortality probability by age matrix.
Their arguments can be added.}

\item{p_offspring:}{Probability of a woman having a son per year. Default is 0.3.}
}
\value{
data.frame with two columns, where the number of rows is the number of
people. The first column contains the ages and the second column contains the sex.
}
\description{
Protocol to generate a stochastic demographic process
}
