% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_simulation.R
\name{initKs}
\alias{initKs}
\title{Initialize carrying capacity}
\usage{
initKs(
  Kbase = c(HG = 30, F = 120),
  sites,
  ressources,
  sizeexp = NULL,
  rate = 0.5
)
}
\arguments{
\item{Kbase}{baseline carrying capacity or different cultures}

\item{sites}{raster with site and culture}

\item{ressources}{a raster type for fall sites}

\item{sizeexp}{Cultures for which to apply a random number generated by an exponential
distribution that multiplies the amount of resources}

\item{rate}{Rate of exponential distribution}
}
\value{
A list with a carrying capacity for each culture
}
\description{
It generates a random carrying capacity for each site, but based on the amount
of resources.
}
