% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_simulation.R
\name{initlistsites}
\alias{initlistsites}
\title{Initialize list of sites}
\usage{
initlistsites(list_sites, ts = 200)
}
\arguments{
\item{list_sites}{a list of population for all initial sites of the simulation}

\item{ts}{length of simulation}
}
\value{
matrix of list sites
}
\description{
Initialize list of sites across the years of a simulation
}
