% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_simulation.R
\name{initpopstruc}
\alias{initpopstruc}
\title{Create initial population matrix}
\usage{
initpopstruc(n = 100, ages = 10:30, p_sex = c(0.5, 0.5))
}
\arguments{
\item{n}{number of individuals}

\item{ages}{initial ages}

\item{p_sex}{proportion individuals of different sex}
}
\value{
a dataframe with age and sex of every individual of the population
}
\description{
This functions initializes a population matrix
}
