% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{K_lim}
\alias{K_lim}
\title{Simulation of carrying capacity limitation}
\usage{
K_lim(x, K, prob = 0.8)
}
\arguments{
\item{x}{Data frame or matrix. Population (number of people)}

\item{K}{Integer. Carrying capacity. Provided by the user}

\item{prob}{It is the probability of dying when surpassing carrying capacity}
}
\value{
Returns a data.frame with the updated population
}
\description{
If the population exceeds the carrying capacity, it eliminates oversize
with prob probability per person exceeding.
}
