% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_decay.R
\name{logisticdecay}
\alias{logisticdecay}
\title{Logistic decay}
\usage{
logisticdecay(pt, rast, L = 1, k = 1e-04, x0 = 60000)
}
\arguments{
\item{pt}{a vector points around witch decay is computed}

\item{rast}{a raster to compute distances}

\item{L}{starting point of decay}

\item{k}{a vector of decay rates}

\item{x0}{areas around \code{pt} in which resources decay from}
}
\value{
Logistic decay vector
}
\description{
Given a raster and a vector of points, set resources hotspots with a
logistic decay around them
}
