% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_loss.R
\name{long_loss}
\alias{long_loss}
\title{Long-term taphonomic loss}
\usage{
long_loss(x, theta_l, it)
}
\arguments{
\item{x:}{A vector with the amount of sample per depth.}

\item{theta_l:}{Probability of the record surviving after the first year after deposition}

\item{it:}{Initial time. Initial year of occupation in BP.}
}
\value{
\code{numeric} with reduced values due to long-term taphonomic loss
}
\description{
Simulates the long-term taphonomic loss applying binomials to values on a
single vector for 1 year. If used for many years, use with apply function,
where the rows of the data frame are the depths and the columns the years.
}
