% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlin_noise.R
\name{perlin_noise}
\alias{perlin_noise}
\title{Perlin-noise for elevation and slope}
\usage{
perlin_noise(n = 5, m = 7, N = 100, M = 100)
}
\arguments{
\item{n}{Size of the grid in the vector field on one dimension}

\item{m}{Size of the grid in the vector field on the other dimension}

\item{N}{Size of the final image}

\item{M}{Size of the final image on the other dimension}
}
\value{
\code{numeric} matrix of size N-1 x M-1
}
\description{
This function creates slope and elevation with 2-D autocorrelation and noise.
Collectively it is called Perlin noise
}
