% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plotMap}
\alias{plotMap}
\title{Plot the elevation and water map of the simulation}
\usage{
plotMap(height, water, maintitle = "")
}
\arguments{
\item{height}{Elevation raster}

\item{water}{Water}

\item{maintitle}{Title of the plot}
}
\description{
Plot the elevation and water map of the simulation
}
