% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{Pop_stoch}
\alias{Pop_stoch}
\title{Stochastic population generation}
\usage{
Pop_stoch(pop_size, K, ts, prob = 0.8, ...)
}
\arguments{
\item{pop_size}{Integer, the initial population}

\item{K}{Only if model_pop = TRUE. In this case, it is the carrying capacity}

\item{ts}{Time-span, the number of years considered for the process}

\item{prob}{Probability that an individual will die if total population
exceeds K. Default is 0.8}

\item{...}{Additional arguments passed to \code{\link[=Gpd]{Gpd()}}, and in turn to \code{\link[=death]{death()}}}
}
\value{
A vector with the population size for each year from 1 to ts
}
\description{
It reproduces the population stochastic process. The result is a vector with the
number of individuals for each year.
}
