% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_deposition.R
\name{Rec_c}
\alias{Rec_c}
\title{Archaeological deposition record}
\usage{
Rec_c(x, area, ts, InitBP, persqm = FALSE, ...)
}
\arguments{
\item{x}{Vector with the number of samples per year. As produced as produced by iterating over A_rates.}

\item{area}{Only if persqm = TRUE. In this case, the total area of the site
must be provided}

\item{ts}{Time-span, the number of years considered for the process}

\item{InitBP}{Initial year considered for the process. In BP.}

\item{persqm}{If TRUE, the total record is divided by the area of the site
(in square meters), so that the output is per square meter. Default is FALSE}

\item{...}{This function uses the functions \code{D_along()}. The additional
arguments can be added.}
}
\value{
A square matrix of size ts. It contains the amount of samples deposited at each year (columns) and each depth (rows)
}
\description{
Generate archaeological deposition record over time and depth
It iterates over D_along to spread the amount of samples produced at each time point over different profundities
}
