% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run simulation}
\usage{
run_simulation(
  cultures = NULL,
  viable = viable,
  sites = sites,
  dem = height.ras,
  ressources = ress,
  water = height.wat,
  foldervid = "pathtofinal",
  visu = FALSE,
  visumin = TRUE,
  ts = 20000,
  Kbase = c(HG = 35, F = 120),
  cul_ext = c(HG = 7, F = 6),
  penal_cul = c(HG = 4, F = 5),
  prob_birth = c(HG = 0.3, F = 0.5),
  prob_survive = c(HG = 0.8, F = 0.6),
  prob_split = c(HG = 0.2, F = 0.6),
  prob_move = c(HG = 0.2, F = 0.1),
  minimals = c(HG = 0.14, F = 0.2),
  bufferatack = 400,
  buffersettl = 2000,
  Nts = NULL,
  Ips = NULL
)
}
\arguments{
\item{cultures}{Vector of cultures to be simulated. Default is \code{NULL},
in which case it is taken from \code{sites}. Culture names must also be present in
as names in the named vectors with the simulation parameters for each culture.}

\item{viable}{Viable SpatVector of the territory}

\item{sites}{SpatVector of sites}

\item{dem}{Digital Elevation Model SpatRaster of the map}

\item{ressources}{SpatVector with resources position}

\item{water}{SpatRaster with water}

\item{foldervid}{If \code{visu=TRUE}, folder to save plots with simulation snaps}

\item{visu}{logical; whether to plot simulation snaps}

\item{visumin}{logical; whether to plot a minimal visualization of the simulation}

\item{ts}{Length of the simulation in years}

\item{Kbase}{Named vector with carrying capacities for the cultures}

\item{cul_ext}{Named vector. Spatial penalty to extent: lower, bigger penality}

\item{penal_cul}{Named vector. Penality of occupational area:
if low, other sites can come close}

\item{prob_birth}{Named vector. Probability of giving birth every year}

\item{prob_survive}{Named vector. Probability of dying when the pop size is greater than the
carrying capacity}

\item{prob_split}{Named vector. Probability of creating a new split settlement when the pop
size is greater than the carrying capacity}

\item{prob_move}{Probability of migrating to a existing settlement when Ne > K}

\item{minimals}{Named vector. How big, proportionally, the group of migrants
should be to create a new city vs migrate to a existing one.}

\item{bufferatack}{Maximum distance around which a settlement can fight}

\item{buffersettl}{Minimum distance around a site in which a new settlement cannnot settle}

\item{Nts}{Initiallized list of sites. Created with \link{initlistsites}.
Default is \code{NULL}, in which case it's created inside the function from \code{sites}.}

\item{Ips}{Initial population structure. Created with \link{initpopstruc}.
Default is \code{NULL}, in which case it's created inside the function from \code{sites}}
}
\value{
A list with site data across the simulation period, population structures, war
casualties and updated sites positions.
}
\description{
This function runs a stochastic simulation in which different cultures
positioned in sites in a raster interact, grow, die, fight, migrate,...
}
