% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_loss.R
\name{short_loss}
\alias{short_loss}
\title{Short-term taphonomic loss}
\usage{
short_loss(x, theta_s)
}
\arguments{
\item{x:}{A vector with the amount of samples per depth, at a specific years.}

\item{theta_s:}{Probability of the record surviving after the
first year after deposition}
}
\value{
\code{numeric} with reduced values due to the short-term taphonomic loss
}
\description{
Simulates the short-term taphonomic loss applying binomials to values on a
single vector for 1 year. If used for many years, use with apply function,
where the rows of the data frame are the depths and the columns the years.
}
