% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{simplefight}
\alias{simplefight}
\title{Model a simple fight}
\usage{
simplefight(Ne, a, b)
}
\arguments{
\item{Ne}{list of population sizes for the fighting settlement}

\item{a}{indice of the first settlement}

\item{b}{indice of the second settlement}
}
\value{
Returns the updated population size of both settlements engaged in the fight
}
\description{
A function to compute lost during a fighting.
The winner and looser are decided probabilistically from their relative size.
It then uses a binomial to model the loss of population sizes,
and assigns a probability of 0.9 for the winner and 0.4 for the looser.
}
