% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{TK95}
\alias{TK95}
\title{Power law Noise from Kimmer & Koening}
\usage{
TK95(N, alpha = 1)
}
\arguments{
\item{N}{the length of the timeserie to generate}

\item{alpha}{the slope of the power distribution (called omega in whitehead)}
}
\description{
From tuneR package (implemented in the file Waveforms.R in tuneR source)

Based on Timmer & Koening (1995)
}
