% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{warpoints}
\alias{warpoints}
\title{Draw a war symbol where two clans are fighting}
\usage{
warpoints(sites, a, b, Ne, buffersize = 300, plot = T, sizewar = 2)
}
\arguments{
\item{sites}{Raster with site coordinates, cultures and carrying capacities}

\item{a}{Index of first settlement}

\item{b}{Index of second settlement}

\item{Ne}{Population sizes list}

\item{buffersize}{buffer size around a given site to consider contacts.
It's a factor that multiplies the population size Ne}

\item{plot}{Whether to make a plot or not}

\item{sizewar}{Size of the war symbol}
}
\value{
Raster of the intersection point between sites a and b, if any.
}
\description{
Draws a war symbol in the sites raster at the point of intersection between
fighting clans
}
