% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{whotouch}
\alias{whotouch}
\title{check who's touchin site i}
\usage{
whotouch(i, sites, homophily = F, buffersize = 200, Ne)
}
\arguments{
\item{i}{indice of the site checked}

\item{sites}{raster with site coordinates,  cultures and karying capacityes}

\item{homophily}{if true, return all sites that touch (even same culture)}

\item{Ne}{population size of all sites (could be remove if Ne was stored in the raster)}
}
\description{
check who's touchin site i
}
