% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{long_loss}
\alias{long_loss}
\title{long_loss
Returns a data frame in the same form that the inserted data frame with
reduced values for short-term taphonomic loss process}
\usage{
long_loss(x, theta_l, it)
}
\arguments{
\item{x:}{A vector with the amount of sample per depth. If used for many years, use
with apply function, where the rows of the data frame are the depths and the columns
the years.}

\item{theta_l:}{Probability of the record surviving after the first year after deposition}

\item{it:}{Initial time. Initial year of occupation in BP.}
}
\description{
long_loss
Returns a data frame in the same form that the inserted data frame with
reduced values for short-term taphonomic loss process
}
