% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{short_loss}
\alias{short_loss}
\title{short_loss
Returns a data frame in the same form that the inserted data frame with
reduced values for short-term taphonomic loss process}
\usage{
short_loss(x, theta_s)
}
\arguments{
\item{x:}{A vector with the amount of sample per depth. If used for many years, use
with apply function, where the rows of the data frame are the depths and the columns
the years.}

\item{theta_s:}{Probability of the record surviving after the first year after deposition}
}
\description{
short_loss
Returns a data frame in the same form that the inserted data frame with
reduced values for short-term taphonomic loss process
}
