% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{whotouch}
\alias{whotouch}
\title{Check sites touching}
\usage{
whotouch(i, sites, Ne, homophily = F, buffersize = 200)
}
\arguments{
\item{i}{Index of the site checked}

\item{sites}{Raster with site coordinates, cultures and carrying capacities}

\item{Ne}{population size of all sites}

\item{homophily}{if true, return all sites that touch (even same culture)}

\item{buffersize}{Buffer size around a given site to consider contacts.
It's a factor that multiplies the population size Ne}
}
\value{
Returns a raster with the sites that are touching site \code{i}
}
\description{
Check who's touching a given site
}
